requirements.txt
chainlit
langchain
python-dotenv
